rename plat_control to plat_hwm
