+++ meta-facebook/aalc_predev/src/platform/plat_modbus.h
@@ -40,28 +40,6 @@
 int init_custom_modbus_server(void);
 void init_modbus_command_table(void);
 
-enum PUMP_STATE {
-	PUMP_REDUNDENT_SWITCHED = 0,
-	MANUAL_CONTROL_PUMP = 1,
-	MANUAL_CONTROL_FAN = 2,
-	AUTOTUNE_FLOW_CONTROL = 3,
-	AUTOTUNE_PRESSURE_BALANCE_CONTROL = 4,
-	SYSTEM_STOP = 5,
-	RPU_REMOTE_POWER_CYCLE = 6,
-	MANUAL_CONTROL = 9,
-	CLEAR_PUMP_RUNNING_TIME = 10,
-	CLEAR_LOG = 11,
-	PUMP_1_RESET = 12,
-	PUMP_2_RESET = 13,
-	PUMP_3_RESET = 14,
-};
-
-typedef struct _pump_reset_struct {
-	uint8_t function_index;
-	bool (*fn)(struct _pump_reset_struct *, uint8_t bit_value);
-	uint8_t senser_num;
-} pump_reset_struct;
-
 typedef struct _modbus_command_mapping {
 	uint16_t addr;
 	uint8_t (*wr_fn)(struct _modbus_command_mapping *);
